backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
pymongo
python-dotenv
